Clean up flash library spacing and organize flash tests
