+++ assert_tests/mem_byte_write_test/makefile
@@ -1,3 +1,3 @@
-PROG = flash
+PROG = mem_byte_write_test
 SRC = $(addprefix ../../src/,mem.c)
 include ../makefile
