+++ assert_tests/mem_multibyte_test/makefile
@@ -1,3 +1,3 @@
-PROG = mem_multibyte
+PROG = mem_multibyte_test
 SRC = $(addprefix ../../src/,mem.c)
 include ../makefile
