+++ assert_tests/mem_test/makefile
@@ -1,3 +1,3 @@
-PROG = mem_byte_write
+PROG = mem_test
 SRC = $(addprefix ../../src/,mem.c)
 include ../makefile
