+++ assert_tests/mem_test/mem_test.c
@@ -1,21 +1,21 @@
-// standard C ;ibraries
-#include  	<stdbool.h>
-#include  	<stdint.h>
+// standard C libraries
+#include    <stdbool.h>
+#include    <stdint.h>
 
 // avr-libc includes
-#include 	<avr/io.h>
-#include 	<util/delay.h>
+#include    <avr/io.h>
+#include    <util/delay.h>
 
 // lib-common includes
-#include 	<spi/spi.h>
-#include 	<uart/uart.h>
-#include 	<uart/log.h>
-#include 	<can/can.h>
+#include    <spi/spi.h>
+#include    <uart/uart.h>
+#include    <uart/log.h>
+#include    <can/can.h>
 
 // project file includes
 
-#include 	"../../src/mem.h"
-#include 	"../../src/rtc.c"
+#include     "../../src/mem.h"
+#include     "../../src/rtc.c"
 
 int main(void) {
     init_uart();
@@ -27,12 +27,12 @@ int main(void) {
     _delay_ms (5000);
 
     /*
-      What should happen in this test:
-        1. Initializes all stacks- can be checked with those EEPROM addresses
-          print statements
-        2. Write values: 0x08, 0x07, 0x06, 0x08 into SCI STACK
-        3. Read back the same values from SCI STACK- there should be a header
-        FFF and then these 4 values.
+    What should happen in this test:
+    1. Initializes all stacks- can be checked with those EEPROM addresses
+    print statements
+    2. Write values: 0x08, 0x07, 0x06, 0x08 into SCI STACK
+    3. Read back the same values from SCI STACK- there should be a header
+    FFF and then these 4 values.
     */
     uint8_t i;
 
@@ -46,17 +46,17 @@ int main(void) {
     for (j=0;j<4;j++) {
         uint8_t read_test [10] = {0};
         uint8_t test [4]= {0x08, 0x07, 0x06, 0x08};
-      /*  for(i=0; i<4; i++) {
-          print ("%x\n", test[i]);
+        /*  for(i=0; i<4; i++) {
+        print ("%x\n", test[i]);
         }*/
         write_to_flash(SCI_TYPE, j, test);
         print ("\n***********Read back from expected address: %x******\n", 0x250+8+SCI_INIT+(4*j));
         mem_read(0x250+8+(SCI_INIT+(4*j)), &read_test, 10);
         for(i=0; i<10; i++) {
-          print ("%x\n", read_test[i]);
+            print ("%x\n", read_test[i]);
         }
 
-  }
+    }
 
     print ("*** End of test******\n");
